Fixed Yoad's comments, created useCompletedCheckerEffect hook and  created Nullable generic type
